--- src/main/java/com/sparta/spartaoutsourcing/domian/auth/config/PasswordConfig.java
@@ -1,15 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.auth.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class PasswordConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
