--- src/main/java/com/sparta/spartaoutsourcing/domian/auth/config/WebSecurityConfig.java
@@ -1,112 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.auth.config;
-
-import com.sparta.spartaoutsourcing.domian.auth.UserLogoutHandler;
-import com.sparta.spartaoutsourcing.domian.auth.jwt.JwtAuthenticationFilter;
-import com.sparta.spartaoutsourcing.domian.auth.jwt.JwtAuthorizationFilter;
-import com.sparta.spartaoutsourcing.domian.auth.jwt.JwtUtil;
-import com.sparta.spartaoutsourcing.domian.auth.security.UserDetailsServiceImpl;
-import com.sparta.spartaoutsourcing.domian.auth.token.TokenBlacklistService;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.authentication.logout.LogoutFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-@Slf4j(topic = "WebSecurityConfig")
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity(securedEnabled = true)
-@RequiredArgsConstructor
-public class WebSecurityConfig {
-    private final JwtUtil jwtUtil;
-    private final UserDetailsServiceImpl userDetailsService;
-    private final TokenBlacklistService tokenBlacklistService;
-    private final AuthenticationConfiguration authenticationConfiguration;
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
-        return configuration.getAuthenticationManager();
-    }
-
-    @Bean
-    public JwtAuthenticationFilter jwtAuthenticationFilter() throws Exception {
-        JwtAuthenticationFilter filter = new JwtAuthenticationFilter(jwtUtil);
-        filter.setAuthenticationManager(this.authenticationManager(authenticationConfiguration));
-        return filter;
-    }
-
-    @Bean
-    public JwtAuthorizationFilter jwtAuthorizationFilter() {
-        return new JwtAuthorizationFilter(jwtUtil, userDetailsService, tokenBlacklistService);
-    }
-
-    @Bean
-    public AuthenticationEntryPoint unauthorizedEntryPoint() {
-        return (request, response, authException) -> {
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            response.getWriter().write(authException.getMessage());
-        };
-    }
-
-    @Bean
-    public AccessDeniedHandler accessDeniedHandler() {
-        return (request, response, accessDeniedException) -> {
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            response.getWriter().write(accessDeniedException.getMessage());
-        };
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity.csrf(csrf -> csrf.disable());
-
-        httpSecurity.sessionManagement(sessionManagement ->
-                sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        );
-
-        httpSecurity.authorizeHttpRequests(authorizeHttpRequests ->
-                authorizeHttpRequests
-                        .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll()
-                        .requestMatchers("/users/**").permitAll()
-                        .requestMatchers("/error").permitAll()
-                        .anyRequest().authenticated()
-        );
-
-        httpSecurity.addFilterBefore(jwtAuthorizationFilter(), LogoutFilter.class)
-                .logout(logout ->
-                    logout
-                        .logoutRequestMatcher(new AntPathRequestMatcher("/users/logout", "POST"))
-                        .addLogoutHandler(new UserLogoutHandler(tokenBlacklistService))
-                        .logoutSuccessHandler((request, response, authentication) -> {
-                                response.setStatus(HttpServletResponse.SC_OK);
-                                response.getWriter().write("Logout Success");
-                        })
-                        .logoutSuccessUrl("/")
-                        .invalidateHttpSession(true)
-        );
-
-        httpSecurity.exceptionHandling(handler ->
-                handler
-                    .authenticationEntryPoint(unauthorizedEntryPoint())
-                    .accessDeniedHandler(accessDeniedHandler())
-        );
-
-        httpSecurity.addFilterBefore(jwtAuthorizationFilter(), JwtAuthenticationFilter.class);
-        httpSecurity.addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
-
-        return httpSecurity.build();
-    }
- }
