--- src/main/java/com/sparta/spartaoutsourcing/domian/store/controller/FavoritesController.java
@@ -1,4 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.store.controller;
-
-public class FavoritesController {
-}
