--- src/main/java/com/sparta/spartaoutsourcing/domian/store/exception/FavoriteDuplicationException.java
@@ -1,8 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.store.exception;
-
-public class FavoriteDuplicationException extends IllegalArgumentException{
-
-    public FavoriteDuplicationException(String message) {
-        super(message);
-    }
-}
