--- src/main/java/com/sparta/spartaoutsourcing/domian/store/exception/MaxStoreLimitException.java
@@ -1,7 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.store.exception;
-
-public class MaxStoreLimitException extends RuntimeException{
-    public MaxStoreLimitException(String message) {
-        super(message);
-    }
-}
