--- src/main/java/com/sparta/spartaoutsourcing/domian/store/exception/StoreNotFoundException.java
@@ -1,7 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.store.exception;
-
-public class StoreNotFoundException extends NullPointerException {
-    public StoreNotFoundException(String message) {
-        super(message);
-    }
-}
