--- src/main/java/com/sparta/spartaoutsourcing/domian/store/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.sparta.spartaoutsourcing.domian.store.exception;
-
-public class UserNotFoundException extends NullPointerException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
