+++ src/main/java/com/sparta/spartaoutsourcing/store/controller/FavoritesController.java
@@ -1,0 +1,4 @@
+package com.sparta.spartaoutsourcing.store.controller;
+
+public class FavoritesController {
+}
