+++ src/main/java/com/sparta/spartaoutsourcing/store/exception/FavoriteDuplicationException.java
@@ -1,0 +1,8 @@
+package com.sparta.spartaoutsourcing.store.exception;
+
+public class FavoriteDuplicationException extends IllegalArgumentException{
+
+    public FavoriteDuplicationException(String message) {
+        super(message);
+    }
+}
