+++ src/main/java/com/sparta/spartaoutsourcing/store/exception/MaxStoreLimitException.java
@@ -1,0 +1,7 @@
+package com.sparta.spartaoutsourcing.store.exception;
+
+public class MaxStoreLimitException extends RuntimeException{
+    public MaxStoreLimitException(String message) {
+        super(message);
+    }
+}
