+++ src/main/java/com/sparta/spartaoutsourcing/store/exception/StoreNotFoundException.java
@@ -1,0 +1,7 @@
+package com.sparta.spartaoutsourcing.store.exception;
+
+public class StoreNotFoundException extends NullPointerException {
+    public StoreNotFoundException(String message) {
+        super(message);
+    }
+}
