+++ src/main/java/com/sparta/spartaoutsourcing/store/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.sparta.spartaoutsourcing.store.exception;
+
+public class UserNotFoundException extends NullPointerException {
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}
